Development feature playtester patches

+++ Assets/Scripts/Prototype 2/Components/Bounces.cs
@@ -38,7 +38,7 @@ public class Bounces : MonoBehaviour
         onBouncesUpdated.Invoke();
     }
 
-    public float GetCurrentBounces()
+    public int GetCurrentBounces()
     {
         return bounces;
     }
